pyproject.toml
[project]
name = "crew_azam"
version = "0.1.0"
description = "crew-azam using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.12.2",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
]

[project.scripts]
crew_azam = "crew_azam.main:run"
run_crew = "crew_azam.main:run"
train = "crew_azam.main:train"
replay = "crew_azam.main:replay"
test = "crew_azam.main:test"
run_with_trigger = "crew_azam.main:run_with_trigger"
scan_unread_emails = "crew_azam.main:scan_unread_emails"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
